// src/app/Modules/shared/models/Data/Reembolso.ts
// Generated by https://quicktype.io

export interface Reembolso {
  reembolso_id: number;
  siniestro_id:              number;
  cliente:                   number;
  nombre_propietario_cuenta: string;
  nro_identificacion:        string;
  nro_cuenta:                string;
  codigo_swift:              string;
  banco:                     string;
  pais:                      string;
  ciudad:                    string;
  fecha_emision:             string;
  status:                    number;
}




export interface ReembolsoPost extends Omit<Reembolso, 'reembolso_id'>{
}